fix: switch to `npm run type-check` for type-checking hook
The `tsc` command needs a `tsconfig.json` file to be present in the
working directory, but in a monorepo root there is no such a file.
Use `npm run type-check` instead, which runs `tsc` with `tsconfig.json`
of individual packages and fixes the issue.

Also remove `typescript` dependency from the monorepo root, as it is no
longer needed.