fix: fix vitest hangup
Because of a bug in vitest, it doesn't work correctly in multi-threaded
mode on some operating systems (only in some situations).

Disable threads to solve the issue.

Please note that the usage of `singleThread: true` in vitest config is
discouraged, as it runs all of the tests in the same environment, unlike
`threads: false` which runs each test in a child process.

Link to vitest issue (open at the time of commit):
https://github.com/vitest-dev/vitest/issues/2008