fix: switch to node loaders instead of experimental flag
For node versions newer than 18, the experimental specifier resolution
flag is removed in favor of newer loader flag. Switch to the new flag,
adding it to `start`, `start:prod`, and `test` npm scripts.

The `extensionless` loader is used to mimic previous experimental flag
behavior (for loading files without an extension).

In addition, change all coverage npm scripts to run the test script with
additional coverage flag. The point of this is, if some config, etc. is
used for the test script, the same applies to coverage without the need
for additional changes.